Sometimes the getOWLOntologyByURI fails without throwing an exception.  If we want getDefaultOWLOntology to be the default we should wait for the end before checking if we failed.
I am not at all sure this is the right thing to do here.  This change was triggered by another bug which I am going to ignore for the nonce.  The ontology had a name of the form "file:///C:/...".   The URIUtilities method replaced this with a name of the form "file:/C:/..." which was not found as an ontology.

git-svn-id: http://smi-protege.stanford.edu/repos/protege/owl/trunk@21163 30a32298-b1ee-0310-87e3-a7336bc3aa68